Add Coverage and Debug test launchers.
